week3/requirements.txt
fastmcp>=0.1.0
httpx>=0.25.0
anthropic>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
# Honeycomb.io - Optional (install with: pip install opentelemetry-api opentelemetry-sdk opentelemetry-exporter-otlp opentelemetry-instrumentation-httpx opentelemetry-instrumentation-fastapi)
# opentelemetry-api
# opentelemetry-sdk
# opentelemetry-exporter-otlp
# opentelemetry-instrumentation-httpx
# opentelemetry-instrumentation-fastapi
